requirements.txt
torch 
torchvision 
torchaudio
transformers
opencv-python 
scikit-image 
albumentations
pandas 
numpy 
scikit-learn 
matplotlib
faiss-cpu
SimpleITK
tqdm
streamlit
seaborn
grad-cam
kagglehub
gdown